+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Added
 - registry/api/v2: Show version and supported extra features in custom headers
 
+## Changed
+- registry/handlers: Encapsulate the value of err.detail in logs in a JSON object
+
 ### Fixed
 - registry/storage: Fix panic during uploads purge